+++ HISTORY.md
@@ -1,3 +1,7 @@
+# Version 0.0.8
+- Simplified and extended `calculate_residuals_statistics()`
+- Added covariance type and fixed effects table to `linearmodels.panel.result.PanelEffectsresult` print result
+
 # Version 0.0.7
 - Moved type checks to functions and globals
 - Removed unnecessary whitespace from model formula